Add game-over message
When a player reaches the target score,
a win message is logged to the console.
A "new round" button has been added,
in order to avoid triggering a prompt upon
loading the page.  The button won't work after
the game is over. The "board" variable
previously in "gameController" has been moved
inside "playRound", so that whenever a new
round is started, the factory function
stored in "board" will create a new board
from scratch