Make user mark his sign directly on board UI
Instead of writing in a prompt the coordinates,
the user will now have to click the according
cell, which will be marked with his sign.
Fixed a bug that would cause the game to freeze
whenever no more moves were available for either
user or cpu. Also made it so that when a round
ends, the first player to make his move
in the subsequent round depends on who made the
last move in the previous one, as it will change
accordingly