Add "restart the game" button & fix checkWin
In place of the old "new round" button, it's
purpose is to clear the existing board out of
all current values/signs to start a new game.
In addition, the last player that has made a move
in the previous round, is to first one to make it
in the new one.
Fixed a bug in checkWin where it would
give false wins under certain conditions