Add result to DOM after game is over
Result now appears under the "restart the game"
button after a game's over, with a message that
either shows the winner or that the game's a tie.
Changed the "checkCells" method to "checkTie"
and moved it from the "Gameboard" to
"gameController" object to fit better within the
code logic. Result message was previously shown in
the console thanks to "checkWin" and "checkTie",
now they only end the game whenever the respective
conditions have been met, while
it's the "winnerScreen" function that takes
care of showing result by adding it to the DOM